Move the metrics to own package and add "hack" counter
Had to move the metrics to its own folder, since it would be used both in the controller package and the logic package. and if it was not split up. then we would have cyclic dependencies.
